+++ R/shorthand.R
@@ -2,7 +2,6 @@
 #' @description see \link[glue]{glue} documentation
 #' @param ... parameters passed to \code{glue()}
 #'
-#' @importFrom glue glue
 #' @return a glue character string
 #' @export
 gg <- function(...){
@@ -14,7 +13,6 @@ gg <- function(...){
 #' @description see \link[dplyr]{glimpse} documentation
 #' @param ... parameters passed to \code{glimpse()}
 #'
-#' @importFrom dplyr glimpse
 #' @return x original x is (invisibly) returned, allowing \code{glimpse()} to be used within a data pipe line.
 #' @export
 g <- function(...){
@@ -26,7 +24,6 @@ g <- function(...){
 #' @description see \link[here]{here} documentation
 #' @param ... parameters passed to \code{here()}
 #'
-#' @importFrom here here
 #' @return root path to project plus any additional path components provided
 #' @export
 hh <- function(...){